+++ webui/media.py
@@ -13,8 +13,16 @@ import torch
 import uuid
 
 
-# Function: process_submission
 def process_submission(transcription_id, upload_url, diarize):
+   """
+   Processes a transcription submission by downloading media, transcribing it, and
+   optionally diarizing the transcription.
+
+   Args:
+      transcription_id (int): The ID of the transcription to process.
+      upload_url (str): The URL from which to download the media file (optional).
+      diarize (bool): Whether to perform diarization on the transcription.
+   """
    try:
       transcription = Transcription.objects.get(pk=transcription_id)
    except Transcription.DoesNotExist:
@@ -43,20 +51,28 @@ def process_submission(transcription_id, upload_url, diarize):
          transcription.fail_incomplete_statuses('Diarizing media failed.')
          return
 
-   return
 
-
-# Function: download_media
 def download_media(transcription_id, upload_url):
+   """
+   This function retrieves a transcription object by its ID, downloads the media file
+   from the provided URL using yt-dlp, and saves the downloaded file to the
+   transcription record.
+
+   Args:
+      transcription_id (int): The ID of the transcription object to associate the media
+         with.
+      upload_url (str): The URL of the media to be downloaded.
+   """
    try:
       transcription = Transcription.objects.get(pk=transcription_id)
    except Transcription.DoesNotExist:
-      return None
+      return
 
-   status = transcription.statuses.get(process = TranscriptionStatus.DOWNLOADING)
-   status.status = TranscriptionStatus.PROCESSING
-   status.start_time = datetime.now()
-   status.save()
+   download_status = transcription.statuses.get(process=TranscriptionStatus.DOWNLOADING)
+   if download_status.status == TranscriptionStatus.FAILED: return
+   download_status.status = TranscriptionStatus.PROCESSING
+   download_status.start_time = datetime.now()
+   download_status.save()
 
    # Can the opts for yt-dlp use a function to generate hex codes on the fly?
    hex = '_' + uuid.uuid4().hex[:7]
@@ -91,15 +107,26 @@ def download_media(transcription_id, upload_url):
    # Delete temp file
    Path(file_path).unlink(True)
 
-   status.status = TranscriptionStatus.COMPLETED
-   status.end_time = datetime.now()
-   status.save()
-
-   return transcription
+   download_status.status = TranscriptionStatus.COMPLETED
+   download_status.end_time = datetime.now()
+   download_status.save()
 
 
-# Function: resegment_words
 def resegment_word_list(word_list, max_characters, max_time):
+   """
+   Resegments a list of words into segments based on character length, time duration,
+   and speaker consistency.
+
+   Args:
+      word_list (list of dict): A list of words to resegment with word, speaker, start,
+         and end attributes.
+      max_characters (int): The maximum number of characters allowed in a segment.
+      max_time (float): The maximum time duration (in seconds) allowed for a segment.
+
+   Returns:
+      list of dict: A list of segments, where each segment is represented as a dictionary.
+   """
+
    # Better than param defaults as checks for ''
    # Segments will get the first word, do not need to check for minimum length or time
    if not max_characters: max_characters = settings.MAX_SEGMENT_LENGTH
@@ -142,17 +169,25 @@ def resegment_word_list(word_list, max_characters, max_time):
    return segments
 
 
-# Function: transcribe_file
 def transcribe_file(transcription_id):
+   """
+   This function retrieves a transcription object from the database, processes the
+   associated audio file using a Whisper model, and saves the transcription results,
+   including word-level timestamps and segmented text, back to the database.
+
+   Args:
+      transcription_id (int): The ID of the transcription object to process.
+   """
    try:
       transcription = Transcription.objects.get(pk=transcription_id)
    except Transcription.DoesNotExist:
-      return None
+      return
 
-   status = transcription.statuses.get(process = TranscriptionStatus.TRANSCRIBING)
-   status.status = TranscriptionStatus.PROCESSING
-   status.start_time = datetime.now()
-   status.save()
+   transcription_status = transcription.statuses.get(process=TranscriptionStatus.TRANSCRIBING)
+   if transcription_status.status == TranscriptionStatus.FAILED: return
+   transcription_status.status = TranscriptionStatus.PROCESSING
+   transcription_status.start_time = datetime.now()
+   transcription_status.save()
 
    DESCRIPTION_MAX_LENGTH = 100
    word_list = []
@@ -211,104 +246,130 @@ def transcribe_file(transcription_id):
    transcription.description = description[:DESCRIPTION_MAX_LENGTH].strip() + '...'
    transcription.save(update_fields=['description'])
 
-   status.status = TranscriptionStatus.COMPLETED
-   status.end_time = datetime.now()
-   status.save()
+   transcription_status.status = TranscriptionStatus.COMPLETED
+   transcription_status.end_time = datetime.now()
+   transcription_status.save()
 
 
-# Function: diarize_separate_overlaps
 def diarize_separate_overlaps(diarization):
+   """
+   Adjusts a list of speaker diarization segments to handle overlapping ranges by
+   splitting them into non-overlapping segments. Overlapping portions are marked with a
+   special 'OVERLAP' speaker.
+
+   Args:
+      diarization (list of dict): A list of dictionaries representing speaker segments.
+
+   Returns:
+      list of dict: A list of dictionaries representing the adjusted speaker segments.
+   """
+   if not diarization:
+      return []
+
    # Sort the speaker ranges by their start time
-   diarization = sorted(diarization, key=lambda x: x['start'])
+   sorted_diarization = sorted(diarization, key=lambda x: x['start'])
+   separated_segments = []
 
-   # Initialize an empty list to store the separated ranges
-   final_segments = []
-
-   # Iterate through the speaker ranges
-   for i in range(len(diarization)):
-         # If this is the first range, add it to the separated ranges list
-
-         if i == 0:
-               final_segments.append(diarization[i])
-         else:
-               # Get the previous range
-               previous_segment = dict(final_segments[-1])
-
-               # If the current range starts after the previous range ends, add it to the separated ranges list
-               if diarization[i]['start'] >= previous_segment['end'] or diarization[i]['speaker'] == previous_segment['speaker']:
-                     final_segments.append(diarization[i])
-               else:
-                     # Otherwise, there is an overlap, so split the ranges
-                     # First, add the part of the previous range that doesn't overlap
-                     final_segments[-1]['end'] = diarization[i]['start']
-                     # Then add the overlap as a new range
-                     overlap = {
-                        'speaker': 'OVERLAP',
-                        'start': diarization[i]['start'],
-                        'end': min(diarization[i]['end'], previous_segment['end'])
-                     }
-                     final_segments.append(overlap)
-                     # Finally, add the part of the current range that doesn't overlap
-                     if diarization[i]['end'] > previous_segment['end']:
-                        non_overlap = {
-                           'speaker': diarization[i]['speaker'],
-                           'start': overlap['end'],
-                           'end': diarization[i]['end']
-                        }
-                     else:
-                        non_overlap = {
-                           'speaker': previous_segment['speaker'],
-                           'start': overlap['end'],
-                           'end': previous_segment['end']
-                        }
-                     final_segments.append(non_overlap)
-
-   return final_segments
-
-
-# Function: diarize_assign_speakers
-def diarize_assign_speakers(transcription):
-   word_list = []
-   if not transcription: return word_list
-   speaker_buckets = diarize_separate_overlaps(transcription.diarization)
-   transcription.refresh_from_db()
-   word_list = transcription.word_list
-   word_list = sorted(word_list, key=lambda x: x['start'])
-   bucket_iter = iter(speaker_buckets)
-   speaker_bucket = next(bucket_iter, None)
-   last_speaker = speaker_bucket.get('speaker', '') if speaker_bucket else ''
-   word_list_index = 0
-
-   # Assign speakers to words in the word_list
-   while word_list_index < len(word_list):
-      word = word_list[word_list_index]
-
-      # Words are left after buckets are exhausted
-      if not speaker_bucket:
-         word['speaker'] = last_speaker
-      elif word['start'] < speaker_bucket['end']:
-         word['speaker'] = speaker_bucket['speaker']
-      else:
-         last_speaker = speaker_bucket['speaker']
-         speaker_bucket = next(bucket_iter)
+   for current_segment in sorted_diarization:
+      if not separated_segments:
+         separated_segments.append(current_segment)
          continue
 
-      word_list_index += 1
+      last_segment = separated_segments[-1]
+
+       # No overlap or same speaker
+      if current_segment['start'] >= last_segment['end'] or current_segment['speaker'] == last_segment['speaker']:
+         separated_segments.append(current_segment)
+      else:
+         # There's an overlap, split the segments
+         # Add the non-overlapping part of the last segment
+         original_end_time = last_segment['end']
+         last_segment['end'] = current_segment['start']
+
+         # Add the overlapping segment
+         overlap_segment = {
+            'speaker': 'OVERLAP',
+            'start': current_segment['start'],
+            'end': min(current_segment['end'], original_end_time)
+         }
+         separated_segments.append(overlap_segment)
+
+         # Add the remaining part of the segment that doesn't overlap
+         if current_segment['end'] > original_end_time:
+            remaining_segment = {
+                  'speaker': current_segment['speaker'],
+                  'start': original_end_time,
+                  'end': current_segment['end']
+            }
+         else:
+            remaining_segment = {
+                  'speaker': last_segment['speaker'],
+                  'start': current_segment['end'],
+                  'end': original_end_time
+            }
+         separated_segments.append(remaining_segment)
+
+   return separated_segments
+
+
+def diarize_assign_speakers(transcription_id):
+   """
+   This function processes a transcription object's word list and diarization data. It
+   assigns a speaker label to each word in the word list based on the diarization
+   information, ensuring that each word is associated with the correct speaker.
+
+   Args:
+      transcription_id (int): The ID of the transcription to be diarized.
+
+   Returns:
+      list: A list of words with updated dictionaries, where each word dictionary
+         includes a 'speaker' key indicating the assigned speaker label.
+   """
+   try:
+      transcription = Transcription.objects.get(pk=transcription_id)
+   except Transcription.DoesNotExist:
+      return
+
+   speaker_buckets = diarize_separate_overlaps(transcription.diarization)
+   word_list = sorted(transcription.word_list, key=lambda x: x['start'])
+   bucket_index = 0
+   last_speaker = ''
+
+   for word in word_list:
+      # Move to the next speaker bucket if the current word's start time exceeds the bucket's end time.
+      while bucket_index < len(speaker_buckets) and word['start'] >= speaker_buckets[bucket_index]['end']:
+         last_speaker = speaker_buckets[bucket_index]['speaker']
+         bucket_index += 1
+
+      # Assign the speaker from the current bucket.
+      if bucket_index < len(speaker_buckets):
+         word['speaker'] = speaker_buckets[bucket_index]['speaker']
+      else:
+         word['speaker'] = last_speaker
 
    return word_list
 
 
-# Function: diarize_file
 def diarize_file(transcription_id):
+   """
+   This function retrieves a transcription object, processes the associated audio file
+   to perform speaker diarization, and updates the transcription with the diarization
+   results. It also handles the segmentation of the diarized audio and updates the
+   transcription's segments accordingly.
+
+   Args:
+      transcription_id (int): The ID of the transcription to be diarized.
+   """
    try:
       transcription = Transcription.objects.get(pk=transcription_id)
    except Transcription.DoesNotExist:
-      return None
+      return
 
-   status = transcription.statuses.get(process = TranscriptionStatus.DIARIZING)
-   status.status = TranscriptionStatus.PROCESSING
-   status.start_time = datetime.now()
-   status.save()
+   diarize_status = transcription.statuses.get(process=TranscriptionStatus.DIARIZING)
+   if diarize_status.status == TranscriptionStatus.FAILED: return
+   diarize_status.status = TranscriptionStatus.PROCESSING
+   diarize_status.start_time = datetime.now()
+   diarize_status.save()
 
    result = []
    meta = transcription.meta
@@ -331,7 +392,7 @@ def diarize_file(transcription_id):
    # Delete temp audio file
    temp_audio.unlink()
 
-   word_list = diarize_assign_speakers(transcription)
+   word_list = diarize_assign_speakers(transcription_id)
    diarized_segments = resegment_word_list(word_list, meta['max_segment_length'], meta['max_segment_time'])
    # Remove existing segments
    transcription.segments.all().delete()
@@ -341,6 +402,6 @@ def diarize_file(transcription_id):
       segment = Segment(**diarized_segment)
       segment.save()
 
-   status.status = TranscriptionStatus.COMPLETED
-   status.end_time = datetime.now()
-   status.save()
+   diarize_status.status = TranscriptionStatus.COMPLETED
+   diarize_status.end_time = datetime.now()
+   diarize_status.save()