+++ webui/static/webui/js/edit.js
@@ -133,14 +133,26 @@
       buttons.forEach(button => {
          const buttonType = button.dataset.type;
 
+         // Disable media buttons if no media
+         if(!mediaPlayer) {
+            switch(buttonType) {
+               case 'play':
+               case 'pause':
+               case 'rewind':
+                  button.disabled = true;
+                  break;
+            }
+         }
+
+         // Add button click functions
          button.addEventListener('click', () => {
-            switch (buttonType) {
+            switch(buttonType) {
                case 'play':
                   mediaPlayer.currentTime = segmentTimeToSeconds(startTime.value, false);
                   mediaPlayer.play();
                   break;
                case 'pause':
-                  mediaPlayer.pause();
+                  mediaPlayer.paused ? mediaPlayer.play() : mediaPlayer.pause();
                   break;
                case 'rewind':
                   let currentTime = mediaPlayer.currentTime;